Quotes_iOS app
created new application project. It is completely same as the old one just the name is new. It will be used for further development.
Reason for adding it is because is somehow messed up old project setting (some files are missing) and it is easier to fix it this way, plus i don't like old name anyway :)