Show all types of one category
After taping on some category in catalog view, you navigate to new view which is suppose to show all elements of one category, for example all authors etc.
For now it only shows which category you seleceted. Rest of it needs to be done, and it needs to be fetched from web service.